usersgateway, online/offline statuses of users

+++ src/users/users.service.ts
@@ -1,10 +1,13 @@
-import { Injectable, OnModuleInit } from '@nestjs/common'
+import { Injectable, NotFoundException, OnModuleInit } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 import { Op } from 'sequelize'
 import { CreateUserDto } from './dto/create-user.dto'
 import { User } from './models/users.model'
 import { ConfigService } from '@nestjs/config'
 import { hashSync } from 'bcrypt'
+import { Socket } from 'socket.io'
+import { TokenPayload } from 'src/auth/types/token-payload'
+import { JwtService } from '@nestjs/jwt'
 
 
 @Injectable()
@@ -12,18 +15,21 @@ export class UsersService implements OnModuleInit {
 
     async onModuleInit(): Promise<void> {
         // creates/updates the admin user
-        await this.userRepository.destroy({ where: { isAdmin: true } })
-        await this.userRepository.create({
-            username: this.configService.get('ADMIN_USERNAME'),
-            email: this.configService.get('ADMIN_EMAIL'),
-            password: hashSync(this.configService.get('ADMIN_PASSWORD'), 7),
-            isAdmin: true,
-        })
+        if (this.configService.get('UPDATE_ADMIN') == true) {
+            await this.userRepository.destroy({ where: { isAdmin: true } })
+            await this.userRepository.create({
+                username: this.configService.get('ADMIN_USERNAME'),
+                email: this.configService.get('ADMIN_EMAIL'),
+                password: hashSync(this.configService.get('ADMIN_PASSWORD'), 7),
+                isAdmin: true,
+            })
+        }
     }
 
     constructor(
         @InjectModel(User) private userRepository: typeof User,
         private readonly configService: ConfigService,
+        private readonly jwtService: JwtService,
     ) {}
 
     async getUsers(limit?: number, offset?: number): Promise<User[]> {
@@ -70,9 +76,39 @@ export class UsersService implements OnModuleInit {
         userId: string
     ): Promise<User> {
         const user: User = await this.userRepository.findByPk(userId)
-        if (user)
-            await user.destroy()
+        if (!user)
+            throw new NotFoundException('User not found')
+
+        await user.destroy()
         return user
     }
+
+    async updateUserLastSeenOnline(
+        userId: string
+    ): Promise<User> {
+        const user: User = await this.userRepository.findByPk(userId)
+        if (!user)
+            throw new NotFoundException('User not found')
+
+        user.lastSeen = new Date(Date.now())
+        await user.save()
+        return user
+    }
+
+    async getUserIdFromSocket(
+        socket: Socket
+    ): Promise<string> {
+        const jwtToken: string = socket.handshake.headers.authorization.split(' ')[1]
+        const tokenPayload: TokenPayload = this.jwtService.decode(jwtToken) as TokenPayload
+        return tokenPayload.id
+    }
+
+    // async getUserFromSocket(
+    //     socket: Socket
+    // ): Promise<User> {
+    //     const userId: string = await this.getUserIdFromSocket(socket)
+    //     const user: User = await this.userRepository.findByPk(userId)
+    //     return user
+    // }
 
 }
